Moving App.tsx to routes/App.tsx
